requirements.txt
pyspark
pandas==2.3.0
matplotlib
seaborn
notebook
jupyterlab
pyarrow
ijson
scikit-learn==1.6.1
numpy
pyyaml
mlflow==3.1.0
xgboost
optuna
imblearn
redis==6.2.0